chore(inline): separate popup logic and put into module
